requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
sqlmodel>=0.0.22
pydantic>=2.10.3
python-dotenv>=1.0.1
pytest>=8.3.4
httpx>=0.27.2
alembic>=1.14.0
PyJWT>=2.9.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
SQLAlchemy>=2.0.0
anyio>=3.4.0
typing-extensions>=4.0.0
psycopg2-binary>=2.9.7